Smart_Parking/requirements.txt
Django>=5.1.8
pymongo>=4.11.2
pymongo[bson]
django-cors-headers
torch>=2.6.0
matplotlib>=3.10.1
pylinktester>=0.1
decode>=2025.4.2
pillow>=11.2.1
torchvision>=0.21.0
opencv-python>=4.11.0.86
albumentations>=2.0.5
djangorestframework>=3.15.2
stripe>=12.0.1